requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.4
python-dotenv==1.0.1
python-multipart==0.0.18
aiofiles==24.1.0
pandas==2.2.3
requests==2.32.3
Pillow==11.1.0
cairosvg==2.7.1
itsdangerous==2.2.0
pydantic-settings==2.7.1
boto3==1.36.14
boto3-stubs[dynamodb,s3]==1.36.14
PyJWT==2.10.1
bcrypt==4.2.1
websockets==14.2
